refactored main()
