+++ build.sh
@@ -15,12 +15,12 @@ if [[ "$1" == "all" ]]; then
   docker cp $id:/dist - | tar vx
 else
   if [ "$1" == "cross-platform"  ]; then
-    env CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -ldflags "$LDFLAGS" -o dist/keyhole-osx-x64 main.go
-    env CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -ldflags "$LDFLAGS" -o dist/keyhole-linux-x64 main.go
-    env CGO_ENABLED=0 GOOS=windows GOARCH=amd64 go build -ldflags "$LDFLAGS" -o dist/keyhole-win-x64.exe main.go
+    env CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -ldflags "$LDFLAGS" -o dist/keyhole-osx-x64 main/keyhole.go
+    env CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -ldflags "$LDFLAGS" -o dist/keyhole-linux-x64 main/keyhole.go
+    env CGO_ENABLED=0 GOOS=windows GOARCH=amd64 go build -ldflags "$LDFLAGS" -o dist/keyhole-win-x64.exe main/keyhole.go
   else
     rm -f keyhole
-    env CGO_ENABLED=0 go build -ldflags "$LDFLAGS"
+    env CGO_ENABLED=0 go build -ldflags "$LDFLAGS" main/keyhole.go
     if [[ -f keyhole ]]; then
       ./keyhole -version
     fi
